fixed CoT indicator in SPH/VAB
Adds the GetMaxThrust(), GetCurrentThrust(), and GetEngine() to FSengine, and adds IThrustProvider to the list of classes that FSengine implements.

The reason why the CoT didn't work was because EditorMarker_CoT.cs relies on IThrustProvider to feed it the necessary information. Luckily for us, it loops through part.Modules.Count and looks for ANY part that implements IThrustProvider, so it doesn't look specifically for ModuleEngines.cs. Not only does this mean that this was a somewhat easy fix, it means that potentially ANY mod that has custom engine modules could easily fix the CoT issue if it did the above workaround.

Oh, and there's no additional setup on the part modder's side, I.E no reconfiguring for this new update since nothing needs to be reconfigured.

Anyways, going back on track, FSengine.cs did not implement said interface and thus it fails the if (thrustProvider != null) statement and returned nil.  Implement some IThrustProvider, add three one-line functi