Fixed CoT marker for SPH/VAB
EditorMarker_CoT.cs relies on IThrustProvider to feed it thrust information. FSEngine doesn't implement this, so I added the appropriate functions (GetMaxThrust(),GetCurrentThrust(),GetEngineType()) and implemented IThrustProvider.

EditorMarker_CoT only uses OnCenterOfThrustQuery, which FSEngine already has. Values returned by GetMaxThrust(),GetCurrentThrust() ,and GetEngineType() are only ever used for the game's FX. Because I wanted my changes to only fix bugs without changing the 'feel' of the mod so to speak, I made GetMaxThrust() and GetCurrentThrust() return 0 so that it doesn't add unnecessary particle effects that would feel more at home in a bloody rocket engine rather than a helicopter propeller.
This is easily changeable in the future, if anybody so wishes.

GetEngineType() is only used for vibrations. Since the goal of my changes to this mod is to fix bugs *without* changing anything else, I didn't want the resulting vibration effect that this results in. Theref